requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.36
sqlmodel==0.0.22
psycopg2-binary
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
requests==2.31.0
alembic==1.13.1
pytest==7.4.4
httpx==0.26.0
ruff
black==24.1.1
python-multipart==0.0.6
boto3
botocore
email-validator==2.3.0
bcrypt==4.1.2
passlib==1.7.4
jinja2==3.1.3
pyseto==1.8.5
kafka-python==2.3.0
